fix: move/fix test package
