+++ src/test/java/com/sharetreats01/viber_chatbot/callback/handler/event/ConversationStartedEventHandlerTest.java
@@ -1,0 +1,27 @@
+package com.sharetreats01.viber_chatbot.callback.handler.event;
+
+import com.sharetreats01.viber_chatbot.callback.handler.event.ConversationStartedEventHandler;
+import com.sharetreats01.viber_chatbot.callback.properties.ChatbotProperties;
+import com.sharetreats01.viber_chatbot.botMessage.service.message.MessageService;
+import org.junit.jupiter.api.BeforeEach;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+
+import static org.mockito.MockitoAnnotations.openMocks;
+
+class ConversationStartedEventHandlerTest {
+    @Mock
+    private ChatbotProperties properties;
+
+    @Mock
+    private MessageService messageService;
+
+    @InjectMocks
+    private ConversationStartedEventHandler conversationStartedEventHandler;
+
+
+    @BeforeEach
+    public void init() {
+        openMocks(this);
+    }
+}
