+++ src/test/java/com/sharetreats01/viber_chatbot/user/service/UserServiceImplTest.java
@@ -1,11 +1,11 @@
 package com.sharetreats01.viber_chatbot.user.service;
 
-import com.sharetreats01.viber_chatbot.interaction.dto.callback.UserFactory;
+import com.sharetreats01.viber_chatbot.botMessage.callback.UserFactory;
 import com.sharetreats01.viber_chatbot.callback.dto.request.event.property.User;
 import com.sharetreats01.viber_chatbot.user.entity.UserEntity;
 import com.sharetreats01.viber_chatbot.user.repository.UserRepository;
 import com.sharetreats01.viber_chatbot.client.viber.client.ViberWebClient;
-import com.sharetreats01.viber_chatbot.infra.viber.dto.GetUserDetailsResponseFactory;
+import com.sharetreats01.viber_chatbot.client.viber.dto.GetUserDetailsResponseFactory;
 import com.sharetreats01.viber_chatbot.client.viber.dto.request.GetUserDetailsRequest;
 import com.sharetreats01.viber_chatbot.client.viber.dto.response.GetUserDetailsResponse;
 import org.junit.jupiter.api.BeforeEach;
